+++ examples/app-router/app/image-optimization/page.tsx
@@ -1,0 +1,14 @@
+import Image from "next/image";
+
+export default function ImageOptimization() {
+  return (
+    <div>
+      <Image
+        src="https://open-next.js.org/architecture.png"
+        alt="Open Next architecture"
+        width={300}
+        height={300}
+      />
+    </div>
+  );
+}