+++ examples/app-router/app/page.tsx
@@ -47,6 +47,9 @@ export default function Home() {
         <Nav href={"/revalidate-tag"} title="Revalidate Tag">
           Revalidate Tag
         </Nav>
+        <Nav href={"/image-optimization"} title="Image Optimization">
+          Image Optimization with next/image
+        </Nav>
       </main>
     </>
   );